requirements.txt
beautifulsoup4==4.12.3
discord.py==2.3.2
selenium==4.22.0